Deixar código mais bonitos e resolver bugs de token

+++ controllers/mensagens.js
@@ -6,9 +6,9 @@ const { getConnectedClients } = require('../sockets/webSocketServer')
 
 const getTokenFrom = request => {
     const auth = request.get('authorization')
-    if (auth && auth.startsWith('Bearer ')) {
+    if (auth && auth.startsWith('Bearer '))
         return auth.replace('Bearer ', '')
-    }
+    
     return null
 }
 
@@ -19,20 +19,20 @@ mensagensRouter.get('/', async (request, response) => {
     response.json(mensagens)
 })
 
-mensagensRouter.post('/', async (request, response) => {
-    console.log(request.body)
-    
+mensagensRouter.post('/', async (request, response) => {    
     const { mensagem } = request.body
+    const token = getTokenFrom(request)
 
-    if (!mensagem) {
+    if (!mensagem)
         return response.status(400).end()
-    }
-    
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-	if (!decodedToken.id) {
+
+    if (!token)
+        return response.status(401).end()
+
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+	if (!decodedToken.id)
 		return response.status(401).json({ error: 'Token inválido' })
-	}
-    
+	
     const user = await User.findById(decodedToken.id)
 
     const novaMensagem = new Mensagem({
@@ -51,36 +51,33 @@ mensagensRouter.post('/', async (request, response) => {
 
 mensagensRouter.delete('/:id', async (request, response) => {    
     const mensagem = await Mensagem.findById(request.params.id)
-	if (!mensagem) {
-		return response.status(404).end()
-	}
+	const token = getTokenFrom(request)
     
-    const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-	if (!decodedToken.id) {
+    if (!mensagem)
+		return response.status(404).end()
+	
+    if (!token) 
+        return response.status(401).end()
+    
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+	
+    if (!decodedToken.id)
 		return response.status(401).json({ error: 'Token inválido' })
-	}
-
+	
     const user = await User.findById(decodedToken.id)
 
-    if (mensagem.user.toString() !== user.id) {
+    if (mensagem.user.toString() !== user.id) 
 		return response.status(401).end()
-	}
-
+	
     await Mensagem.findByIdAndDelete(request.params.id)
-	return response.status(204).end()
+	
+    return response.status(204).end()
 })
 
 mensagensRouter.get('/usuariosConectados', async (request, response) => {
     const valores = getConnectedClients()
     
     response.json(valores)
-})
-
-mensagensRouter.post('/reset', async(request, response) => {
-    await Mensagem.deleteMany({})
-
-    console.log('reset')
-    response.status(204).end()
 })
 
 module.exports = mensagensRouter
